+++ contracts/RMRK/interfaces/IRMRKMultiResourceBase.sol
@@ -3,34 +3,34 @@
 pragma solidity ^0.8.0;
 
 interface IRMRKMultiResourceBase {
-    event ResourceSet(uint32 resourceId);
+    event ResourceSet(uint64 resourceId);
 
-    event ResourceAddedToToken(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceAddedToToken(uint256 indexed tokenId, uint64 resourceId);
 
-    event ResourceAccepted(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceAccepted(uint256 indexed tokenId, uint64 resourceId);
 
-    event ResourceRejected(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceRejected(uint256 indexed tokenId, uint64 resourceId);
 
     event ResourcePrioritySet(uint256 indexed tokenId);
 
     event ResourceOverwriteProposed(
         uint256 indexed tokenId,
-        uint32 resourceId,
-        uint32 overwrites
+        uint64 resourceId,
+        uint64 overwrites
     );
 
-    event ResourceOverwritten(uint256 indexed tokenId, uint32 overwritten);
+    event ResourceOverwritten(uint256 indexed tokenId, uint64 overwritten);
 
-    event ResourceCustomDataSet(uint32 resourceId, uint64 customResourceId);
+    event ResourceCustomDataSet(uint64 resourceId, uint128 customResourceId);
 
     event ResourceCustomDataAdded(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     );
 
     event ResourceCustomDataRemoved(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     );
 
     function acceptResource(uint256 tokenId, uint256 index) external;
@@ -43,11 +43,11 @@ interface IRMRKMultiResourceBase {
 
     function getActiveResources(
         uint256 tokenId
-    ) external view returns(uint32[] memory);
+    ) external view returns(uint64[] memory);
 
     function getPendingResources(
         uint256 tokenId
-    ) external view returns(uint32[] memory);
+    ) external view returns(uint64[] memory);
 
     function getActiveResourcePriorities(
         uint256 tokenId
@@ -55,12 +55,12 @@ interface IRMRKMultiResourceBase {
 
     function getResourceOverwrites(
         uint256 tokenId,
-        uint32 resourceId
-    ) external view returns(uint32);
+        uint64 resourceId
+    ) external view returns(uint64);
 
     function getCustomResourceData(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     ) external view returns (bytes memory);
 
     function tokenURI(
